+++ src/models/index.ts
@@ -1,0 +1,4 @@
+export * from "./api";
+export * from "./image";
+export * from "./post";
+export * from "./user";
