+++ src/models/user.ts
@@ -1,0 +1,31 @@
+/**
+ * The User type defines the structure of the user object fetched from an external API.
+ * @see https://jsonplaceholder.typicode.com/users
+ */
+export type User = {
+  id: number;
+  name: string;
+  username: string;
+  email: string;
+  address: Address;
+  phone: string;
+  website: string;
+  company: Company;
+};
+
+export type Address = {
+  street: string;
+  suite: string;
+  city: string;
+  zipcode: string;
+  geo: {
+    lat: string;
+    lng: string;
+  };
+};
+
+export type Company = {
+  name: string;
+  catchPhrase: string;
+  bs: string;
+};
